Return Status count as float64 so in-process consumers see it correctly
The maintenance sensor's queue_count was always reading 0 because its
type assertion `resp["count"].(float64)` was silently failing for
in-process DoCommand calls. The two paths see different types:

  - gRPC consumers (e.g. the webapp): structpb only supports a single
    numeric type (double), so all numbers cross the wire as float64.
    The webapp has always seen the right value.
  - In-process consumers (e.g. the maintenance sensor, which is a
    sibling resource in the same module): the map flows directly
    between resources without any protobuf serialization, so int stays
    int. The maintenance sensor's float64 assertion failed silently,
    leaving queue_count at its zero value, and is_safe was effectively
    `!armMoving && !isBusy` — queue depth was ignored.

Fix it at the source by casting to float64 in Status() so both paths
see the same type. The maintenance sensor's existing `.(float64)`
assertion now works for both paths.

+++ module.go
@@ -308,7 +308,13 @@ func (s *beanjaminCoffee) Status(ctx context.Context) (map[string]interface{}, e
 	}
 	step, _ := s.currentStep.Load().(string)
 	return map[string]interface{}{
-		"count":        len(orders),
+		// count is returned as float64 so in-process callers (e.g. the
+		// maintenance sensor in this same module) see the same type as
+		// gRPC callers — structpb forces all numbers to double on the
+		// wire, but the in-process path bypasses that conversion.
+		// Without the cast, an in-process .(float64) assertion silently
+		// fails and the caller reads 0.
+		"count":        float64(len(orders)),
 		"orders":       orderMaps,
 		"is_paused":    s.paused.Load(),
 		"is_busy":      s.running.Load(),
